+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 
+ * Always use `File.pathSeparator` when passing multiple paths via the `BUILD_PATH` environment variable
  * Fix `Pointer` losing its owner when mistakenly ignoring deallocators for `const` values returned from adapters
  * Remove unnecessary declared `Exception` from `Indexer.close()` signature ([pull #382](https://github.com/bytedeco/javacpp/pull/382))
  * Make sure `Parser` recognizes base classes of `struct` as `public` by default